feat: modify target membership and delete import statement
